+++ tortoise/__init__.py
@@ -1,7 +1,6 @@
 import asyncio
 import importlib
 import json
-import logging
 import os
 import warnings
 from contextvars import ContextVar
@@ -23,11 +22,10 @@ from tortoise.fields.relational import (
     OneToOneFieldInstance,
 )
 from tortoise.filters import get_m2m_filters
+from tortoise.log import logger
 from tortoise.models import Model, ModelMeta
 from tortoise.transactions import current_transaction_map
 from tortoise.utils import generate_schema_for_client
-
-logger = logging.getLogger("tortoise")
 
 
 class Tortoise:
@@ -711,4 +709,4 @@ def run_async(coro: Coroutine) -> None:
         loop.run_until_complete(Tortoise.close_connections())
 
 
-__version__ = "0.17.7"
+__version__ = "0.17.8"