+++ tortoise/backends/base/schema_generator.py
@@ -1,4 +1,3 @@
-import logging
 from hashlib import sha256
 from typing import TYPE_CHECKING, Any, List, Set, Type, cast
 
@@ -12,8 +11,6 @@ if TYPE_CHECKING:  # pragma: nocoverage
     from tortoise.models import Model
 
 # pylint: disable=R0201
-
-logger = logging.getLogger("tortoise")
 
 
 class BaseSchemaGenerator: