+++ tortoise/contrib/sanic/__init__.py
@@ -1,10 +1,10 @@
-import logging
 from types import ModuleType
 from typing import Dict, Iterable, Optional, Union
 
 from sanic import Sanic  # pylint: disable=E0401
 
 from tortoise import Tortoise
+from tortoise.log import logger
 
 
 def register_tortoise(
@@ -80,12 +80,12 @@ def register_tortoise(
     @app.listener("before_server_start")
     async def init_orm(app, loop):  # pylint: disable=W0612
         await Tortoise.init(config=config, config_file=config_file, db_url=db_url, modules=modules)
-        logging.info("Tortoise-ORM started, %s, %s", Tortoise._connections, Tortoise.apps)
+        logger.info("Tortoise-ORM started, %s, %s", Tortoise._connections, Tortoise.apps)
         if generate_schemas:
-            logging.info("Tortoise-ORM generating schema")
+            logger.info("Tortoise-ORM generating schema")
             await Tortoise.generate_schemas()
 
     @app.listener("after_server_stop")
     async def close_orm(app, loop):  # pylint: disable=W0612
         await Tortoise.close_connections()
-        logging.info("Tortoise-ORM shutdown")
+        logger.info("Tortoise-ORM shutdown")