+++ tortoise/utils.py
@@ -1,8 +1,7 @@
-import logging
 import warnings
 from typing import TYPE_CHECKING
 
-logger = logging.getLogger("tortoise")
+from tortoise.log import logger
 
 if TYPE_CHECKING:  # pragma: nocoverage
     from tortoise.backends.base.client import BaseDBAsyncClient